--- be/src/main/java/com/be/auth/dto/AuthDtos.java
@@ -1,25 +1,0 @@
-package com.be.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public class AuthDtos {
-
-    public record RefreshRequest(
-            @NotBlank String refreshToken
-    ) {}
-
-    public record TokenResponse(
-            String accessToken,
-            String refreshToken
-    ) {}
-
-    public record AuthResponse(
-            Long userId,
-            String name,
-            String email
-    ) {}
-
-    public record MessageResponse(
-            String message
-    ) {}
-}
