+++ be/src/main/java/com/be/auth/dto/AuthResponse.java
@@ -1,0 +1,7 @@
+package com.be.auth.dto;
+
+public record AuthResponse(
+        Long userId,
+        String name,
+        String email
+) {}
