+++ be/src/main/java/com/be/auth/dto/KakaoLoginRequest.java
@@ -1,0 +1,7 @@
+package com.be.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record KakaoLoginRequest(
+        @NotBlank String code
+) {}
