+++ be/src/main/java/com/be/auth/dto/LoginResponse.java
@@ -1,0 +1,10 @@
+package com.be.auth.dto;
+
+public record LoginResponse(
+        Long userId,
+        String name,
+        String email,
+        String accessToken,
+        String refreshToken,
+        boolean isNewUser
+) {}
