+++ be/src/main/java/com/be/auth/dto/MessageResponse.java
@@ -1,0 +1,5 @@
+package com.be.auth.dto;
+
+public record MessageResponse(
+        String message
+) {}
