+++ be/src/main/java/com/be/auth/dto/RefreshRequest.java
@@ -1,0 +1,7 @@
+package com.be.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RefreshRequest(
+        @NotBlank String refreshToken
+) {}
