+++ be/src/main/java/com/be/auth/dto/TokenResponse.java
@@ -1,0 +1,6 @@
+package com.be.auth.dto;
+
+public record TokenResponse(
+        String accessToken,
+        String refreshToken
+) {}
